Split dynamic_boundary_conditions into multiple subpackages
